+++ tests/ludwig/combiners/test_combiners.py
@@ -23,7 +23,7 @@ from ludwig.combiners.combiners import (
     TransformerCombiner,
     TransformerCombinerConfig,
 )
-from ludwig.marshmallow.marshmallow_schema_utils import load_config
+from ludwig.marshmallow.utils import load_config
 
 logger = logging.getLogger(__name__)
 logger.setLevel(logging.INFO)
