+++ tests/ludwig/marshmallow/test_marshmallow_misc.py
@@ -4,8 +4,8 @@ from marshmallow.utils import EXCLUDE
 from marshmallow_dataclass import dataclass
 
 import ludwig.combiners.combiners as lcc
-import ludwig.marshmallow.marshmallow_schema_utils as lusutils
-from ludwig.marshmallow.marshmallow_schema_utils import BaseMarshmallowConfig
+import ludwig.marshmallow.utils as lusutils
+from ludwig.marshmallow.utils import BaseMarshmallowConfig
 from ludwig.models.trainer import TrainerConfig
